removed redundant code

+++ src/controller/nodes/NumNode.ts
@@ -10,7 +10,7 @@ export default class NumNode extends _Node{
 
     parse(){
         var s = this.getAndCheckToken("[0-9]+", true);
-        this.number = parseInt(s);
+        this.number = parseFloat(s);
     }
 
     evaluate(){
